Implement addBoxToImage function in addbox.js
Worked on addbox.js to implement the addBoxToImage function, which adds bounding boxes to the provided imageData. The function has been successfully imported into index.js and will be used to add bounding boxes to the images. The current version saves the boxed image to a file, but further work is required to convert it to base64 and make it accessible to the getEachResult function.